refactor: 게임 시작 기능 도메인에서 컨트롤러로 분리
