+++ src/main/java/racingcar/Application.java
@@ -1,10 +1,12 @@
 package racingcar;
 
+import controller.GameController;
+
 public class Application {
 
     public static void main(String[] args) {
-        Game game = new Game();
-        game.start();
+        GameController gameController = new GameController();
+        gameController.start();
     }
 
 }
